+++ Assets/Scripts/BulletType.cs
@@ -6,7 +6,7 @@ public class BulletType : ScriptableObject
     [SerializeField] public float m_Speed;
     //[SerializeField] public GameObject m_Mesh;
     [SerializeField] public Material m_Material;
-    [SerializeField] public int m_damages;
+    [SerializeField] public float m_damages;
     [SerializeField] public float m_expensionRate;
     [SerializeField] public float m_lifeSpan;
     public float fireRate;
